Add prefetch hot chats in hubs tab
