+++ src/pages/[spaceId]/index.tsx
@@ -1,20 +1,10 @@
-import {
-  getLinkedChatIdsForSpaceId,
-  getSpaceIdFromAlias,
-} from '@/constants/chat-room'
+import { getSpaceIdFromAlias } from '@/constants/chat-room'
 import HomePage, { HomePageProps } from '@/modules/chat/HomePage'
-import { prefetchBlockedEntities } from '@/server/moderation'
-import { getPostQuery } from '@/services/api/query'
-import { getCommentIdsQueryKey } from '@/services/subsocial/commentIds'
-import { getChatIdsBySpaceIdQuery } from '@/services/subsocial/posts'
-import { getSpaceBySpaceIdQuery } from '@/services/subsocial/spaces'
-import { getSubsocialApi } from '@/subsocial-query/subsocial/connection'
+import { prefetchChatPreviewsData } from '@/server/chats'
 import { getMainSpaceId } from '@/utils/env/client'
 import { getCommonStaticProps } from '@/utils/page'
 import { validateNumber } from '@/utils/strings'
-import { PostData } from '@subsocial/api/types'
 import { dehydrate, QueryClient } from '@tanstack/react-query'
-import { getPostsFromCache } from '../api/posts'
 import { AppCommonProps } from '../_app'
 
 export const getStaticPaths = async () => {
@@ -39,34 +29,6 @@ function getSpaceIdFromParam(paramSpaceId: string) {
   return spaceId
 }
 
-async function getLastMessages(messageIdsByChatIds: string[][]) {
-  const lastMessageIds = messageIdsByChatIds
-    .map((ids) => ids[ids.length - 1])
-    .filter((id) => !!id)
-
-  let lastMessages: PostData[] = []
-  if (lastMessageIds.length > 0) {
-    lastMessages = await getPostsFromCache(lastMessageIds)
-  }
-  return lastMessages
-}
-async function getChatPreviewsData(chatIds: string[]) {
-  const subsocialApi = await getSubsocialApi()
-
-  const [messageIdsByChatIds, chats] = await Promise.all([
-    Promise.all(
-      chatIds.map((chatId) => {
-        return subsocialApi.blockchain.getReplyIdsByPostId(chatId)
-      })
-    ),
-    getPostsFromCache(chatIds),
-  ] as const)
-
-  const lastMessages = await getLastMessages(messageIdsByChatIds)
-
-  return { chats, lastMessages, messageIdsByChatIds }
-}
-
 export const getStaticProps = getCommonStaticProps<
   HomePageProps & AppCommonProps
 >(
@@ -79,34 +41,7 @@ export const getStaticProps = getCommonStaticProps<
     if (!spaceId) return undefined
 
     try {
-      const subsocialApi = await getSubsocialApi()
-
-      const chatIds = await subsocialApi.blockchain.postIdsBySpaceId(spaceId)
-      const allChatIds = [...chatIds, ...getLinkedChatIdsForSpaceId(spaceId)]
-
-      const [{ lastMessages, chats, messageIdsByChatIds }] = await Promise.all([
-        getChatPreviewsData(allChatIds),
-        prefetchBlockedEntities(queryClient, allChatIds),
-        getSpaceBySpaceIdQuery.fetchQuery(queryClient, spaceId),
-      ] as const)
-
-      getChatIdsBySpaceIdQuery.setQueryData(queryClient, spaceId, {
-        spaceId,
-        chatIds,
-      })
-      messageIdsByChatIds.forEach((messageIds, idx) => {
-        queryClient.setQueryData(
-          getCommentIdsQueryKey(allChatIds[idx]),
-          messageIds ?? null
-        )
-      })
-      ;[...lastMessages, ...chats].forEach((post) => {
-        getPostQuery.setQueryData(
-          queryClient,
-          post.id,
-          JSON.parse(JSON.stringify(post))
-        )
-      })
+      await prefetchChatPreviewsData(queryClient, spaceId)
     } catch (e) {
       console.error('Error fetching for home page: ', e)
     }
