+++ src/pages/hubs.tsx
@@ -2,9 +2,10 @@ import { getLinkedChatIdsForSpaceId } from '@/constants/chat-room'
 import HubsPage from '@/modules/chat/HubsPage'
 import { HubsPageProps } from '@/modules/chat/HubsPage/HubsPage'
 import { AppCommonProps } from '@/pages/_app'
+import { prefetchChatPreviewsData } from '@/server/chats'
 import { getSpaceBySpaceIdQuery } from '@/services/subsocial/spaces'
 import { getSubsocialApi } from '@/subsocial-query/subsocial/connection'
-import { getSpaceIds } from '@/utils/env/client'
+import { getMainSpaceId, getSpaceIds } from '@/utils/env/client'
 import { getCommonStaticProps } from '@/utils/page'
 import { dehydrate, QueryClient } from '@tanstack/react-query'
 
@@ -25,13 +26,14 @@ export const getStaticProps = getCommonStaticProps<
         visibility: 'onlyVisible',
       })
 
-      await Promise.all(
-        hubsData.map(async (hub) => {
+      await Promise.all([
+        prefetchChatPreviewsData(queryClient, getMainSpaceId()),
+        ...hubsData.map(async (hub) => {
           const chatIds = await subsocialApi.blockchain.postIdsBySpaceId(hub.id)
           const linkedChats = getLinkedChatIdsForSpaceId(hub.id)
           hubsChatCount[hub.id] = chatIds.length + linkedChats.length
-        })
-      )
+        }),
+      ])
 
       hubsData.forEach((hub) => {
         getSpaceBySpaceIdQuery.setQueryData(queryClient, hub.id, hub)
